Agrega el componente de agregar al carrito en la home y en el listado de productos.
