+++ src/app/components/agregar-al-carrito/agregar-al-carrito.component.html
@@ -1,5 +1,5 @@
-<span class="sic-com-agregar-al-carrito" *ngIf="authService.isAuthenticated()">
-  <button mat-icon-button color="warn" (click)="agregarAlCarrito(p)">
+<span class="sic-com-agregar-al-carrito" *ngIf="authService.isAuthenticated() && cliente">
+  <button mat-icon-button color="warn" (click)="showDialog($event)">
     <mat-icon>add_shopping_cart</mat-icon>
   </button>
 </span>
