--- src/app/components/agregar-al-carrito/agregarAlCarritoDialog/agregar-al-carrito-dialog.component.scss
@@ -1,11 +1,0 @@
-.sic-com-agregar-al-carrito-dialog {
-
-  .input-cantidad {
-    width: 100px;
-    padding: 0 10px;
-    input {
-      text-align: right;
-    }
-  }
-
-}
