+++ src/app/components/productos-destacados/productos-destacados.component.scss
@@ -25,6 +25,14 @@
     padding-left: .4rem;
 
     .destacado {
+      position: relative;
+      sic-com-agregar-al-carrito {
+        position: absolute;
+        right: 0;
+        top: 0;
+        z-index: 2;
+        background: rgba(255,255,255, 0.7);
+      }
       &.first-loading {
         display: block;
         height: 100%;
