+++ src/app/components/productos/productos.component.html
@@ -7,8 +7,7 @@
   </div>
   <mat-nav-list class="sic-com-listado" disableRipple="true">
     <div *ngFor="let p of productos">
-<!--      <mat-list-item [routerLink]="['/producto/' + p.idProducto]">-->
-      <mat-list-item>
+      <mat-list-item [routerLink]="['/producto/' + p.idProducto]">
         <img mat-list-avatar class="sic-com-avatar" [src]=p.urlImagen>
         <p mat-line class="sic-com-codigo">{{p.codigo}}</p>
         <p mat-line class="sic-com-descripcion"><strong>{{p.descripcion}}</strong></p>
@@ -16,12 +15,7 @@
           <span class="sic-com-precio-lista" [class.bonificado]="estaBonificado(p)">$ {{p.precioLista | number:'1.2-2':'es-AR' }}</span>
           <span class="sic-com-precio-bonificado" *ngIf="estaBonificado(p)">&nbsp;&nbsp;$ {{ p.precioBonificado | number:'1.2-2':'es-AR' }}</span>
         </p>
-        <div fxLayout="row" fxLayout.xs="column">
-          <button mat-icon-button color="primary" [routerLink]="['/producto/' + p.idProducto]">
-            <mat-icon>info</mat-icon>
-          </button>
-          <sic-com-agregar-al-carrito [producto]="p"></sic-com-agregar-al-carrito>
-        </div>
+        <sic-com-agregar-al-carrito [producto]="p" [cliente]="cliente"></sic-com-agregar-al-carrito>
       </mat-list-item>
       <mat-divider></mat-divider>
     </div>
