+++ src/app/components/productos/productos.component.ts
@@ -7,6 +7,7 @@ import {AuthService} from '../../services/auth.service';
 import {Producto} from '../../models/producto';
 import {CarritoCompraService} from '../../services/carrito-compra.service';
 import {Cliente} from '../../models/cliente';
+import {ClientesService} from '../../services/clientes.service';
 
 @Component({
   selector: 'sic-com-productos',
@@ -23,7 +24,8 @@ export class ProductosComponent implements OnInit, OnDestroy {
   buscarProductosSubscription: Subscription;
   cliente: Cliente = null;
 
-  constructor(private productosService: ProductosService,
+  constructor(private clienteService: ClientesService,
+              private productosService: ProductosService,
               private route: ActivatedRoute,
               private avisoService: AvisoService,
               private authService: AuthService,
@@ -42,6 +44,11 @@ export class ProductosComponent implements OnInit, OnDestroy {
         this.pagina = (Number(queryParams['params'].p) - 1) || 0;
         this.productosService.buscarProductos(params['params'].q || '');
       });
+
+    if (this.authService.isAuthenticated()) {
+      this.clienteService.getClienteDelUsuario(this.authService.getLoggedInIdUsuario())
+        .subscribe((c: Cliente) => this.cliente = c);
+    }
   }
 
   ngOnDestroy() {
@@ -82,37 +89,5 @@ export class ProductosComponent implements OnInit, OnDestroy {
   paginaSiguiente() {
     if (this.pagina + 1 >= this.totalPaginas) { return; }
     this.router.navigate(['/productos', { q: this.busquedaCriteria || '' }], { queryParams: { p: this.pagina + 2 } });
-  }
-
-  agregarAlCarrito(p: Producto) {
-    if (!this.authService.isAuthenticated()) {
-      return;
-    }
-
-    /*
-    this.cargandoAlCarrito = true;
-    this.carritoCompraService.agregarQuitarAlPedido(this.producto, this.cantidad)
-      .subscribe(
-        data => {
-          if (this.cliente) {
-            this.carritoCompraService.getCarritoCompra(this.cliente.id_Cliente)
-              .subscribe(
-                (carrito: CarritoCompra) => {
-                  this.carritoCompraService.setCantidadItemsEnCarrito(carrito.cantRenglones);
-                  this.cargandoAlCarrito = false;
-                },
-                err => {
-                  this.avisoService.openSnackBar(err.error, '', 3500);
-                  this.cargandoAlCarrito = false;
-                }
-              );
-          }
-        },
-        err => {
-          this.cargandoAlCarrito = false;
-          this.avisoService.openSnackBar(err.error, '', 3500);
-        }
-      );
-    */
   }
 }
